+++ cardano-api/src/Cardano/Api/SerialiseUsing.hs
@@ -39,7 +39,7 @@ instance SerialiseAsRawBytes a => ToCBOR (UsingRawBytes a) where
 instance SerialiseAsRawBytes a => FromCBOR (UsingRawBytes a) where
     fromCBOR = do
       bs <- fromCBOR
-      case eitherDeserialiseFromRawBytes ttoken bs of
+      case deserialiseFromRawBytes ttoken bs of
         Right x  -> return (UsingRawBytes x)
         Left (SerialiseAsRawBytesError msg) -> fail ("cannot deserialise as a " ++ tname ++ ".  The error was: " ++ msg)
       where
@@ -87,7 +87,7 @@ deserialiseFromRawBytesBase16 ::
   SerialiseAsRawBytes a => ByteString -> Either String (UsingRawBytesHex a)
 deserialiseFromRawBytesBase16 str =
   case Base16.decode str of
-    Right raw -> case eitherDeserialiseFromRawBytes ttoken raw of
+    Right raw -> case deserialiseFromRawBytes ttoken raw of
       Right x  -> Right (UsingRawBytesHex x)
       Left (SerialiseAsRawBytesError msg) -> Left ("cannot deserialise " ++ show str ++ ".  The error was: " <> msg)
     Left msg  -> Left ("invalid hex " ++ show str ++ ", " ++ msg)